# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Threat Modeling Platform API
  description: |
    # World-Class Threat Modeling Platform API
    
    This is a comprehensive API for an enterprise-grade threat modeling platform that provides:
    
    ## 🚀 Key Features
    - **TMAC (Threat Modeling as Code)** - Define threat models as code
    - **AI-Powered Analysis** - Automated threat identification and risk assessment
    - **Multiple Methodologies** - STRIDE, PASTA, LINDDUN, VAST, and more
    - **Enterprise Integration** - SSO, SIEM, DevOps pipeline integration
    - **Real-time Collaboration** - Multi-user editing and review workflows
    - **Comprehensive Reporting** - Executive summaries to technical details
    
    ## 🔒 Security
    - JWT-based authentication with refresh tokens
    - Role-based access control (RBAC)
    - Rate limiting and DDoS protection
    - CORS security with allowlisted origins
    - Input validation and sanitization
    
    ## 📊 Monitoring
    - Health checks for all services
    - Prometheus metrics
    - Distributed tracing
    - Comprehensive logging
    
    ## 🔧 Service Architecture
    - **API Gateway** (Port 3000) - Central routing and security
    - **Auth Service** (Port 3001) - Authentication and authorization
    - **Core Service** (Port 3002) - Threat models and projects
    - **AI Service** (Port 3003) - ML-powered threat analysis
    - **Diagram Service** (Port 3004) - Visual threat modeling
...
    
  version: '1.0.0'
  contact:
...
  license:
...
servers:
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh access token
      description: Generate new access token using refresh token
      operationId: refreshToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
                  description: Valid refresh token
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login endpoint
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for service-to-service authentication
  schemas:
...
    TokenResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Validation failed"
        details:
          type: array
          items:
            type: string
        code:
          type: string
          example: "VALIDATION_ERROR"
        timestamp:
          type: string
          format: date-time
...
  responses:
...
    Unauthorized:
      description: Unauthorized - authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "Authentication required"
            code: "UNAUTHORIZED"
...